libwasi: move fd apis into a separate file
